+++ server/src/models/dto/account/enable_mfa_dto.rs
@@ -1,0 +1,8 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EnableMfaDto {
+    pub secret_key: String,
+    pub code: String,
+}
