+++ server/src/models/dto/account/mod.rs
@@ -1,2 +1,3 @@
 pub mod update_account_info_dto;
 pub mod update_account_password_dto;
+pub mod enable_mfa_dto;
