+++ server/src/models/dto/auth/mod.rs
@@ -5,3 +5,4 @@ pub mod me_response_dto;
 pub mod register_user_dto;
 pub mod revoke_dto;
 pub mod registration_enabled_dto;
+pub mod random_mfa_secret_key_dto;
