+++ server/src/models/dto/auth/random_mfa_secret_key_dto.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RandomMfaSecretKeyDto {
+    pub secret_key: String,
+}
