+++ server/src/tests/fixtures/users.sql
@@ -1,4 +1,4 @@
 INSERT INTO users
-VALUES ('abc', 'alice', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'user', null, null),
-    ('def', 'bob', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'system', null, null),
-    ('ghi', 'charley', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'system', null, 'MU3EEY32LJTXIMKHKQ3TAR2MIJVUIVKM');
+VALUES ('abc', 'alice', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'user', null, null, null),
+    ('def', 'bob', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'system', null, null, null),
+    ('ghi', 'charley', '$01$TvFRrM4FKhYPqh3y6jQVHl4bItDhDX$0daedac96893022e4c737b0affd43ecd915a8cc83203a5b4b571065bb5793d1c$', 'system', null, 'MU3EEY32LJTXIMKHKQ3TAR2MIJVUIVKM', '{abcd1234}');
